% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtools.R
\name{print.dtools_result}
\alias{print.dtools_result}
\title{Print method for dtools results}
\usage{
\method{print}{dtools_result}(x, n = 10, ...)
}
\arguments{
\item{x}{A dtools_result object}

\item{n}{Maximum number of distributions to display (default: 10)}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for dtools results
}
